% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topserver.R
\name{topserver}
\alias{topserver}
\title{Query Paul's topserver page}
\usage{
topserver()
}
\value{
A character vector of information about the servers from the topserver page.
}
\description{
Query Paul's topserver page
}
\examples{
> topserver()
[1] "anaona: Idle time 61.04\% on 40 CPUs over  52.78 days; load 2.05 3/345 14728"  
[2] "anbona: Idle time 30.09\% on 40 CPUs over  52.78 days; load 50.00 51/430 29895"
[3] "ancona: Idle time 85.98\% on 40 CPUs over  52.78 days; load 2.07 3/352 11143"  
[4] "andona: Idle time 46.86\% on 40 CPUs over  52.78 days; load 40.07 41/369 6075" 
[5] "bivona: Idle time 89.83\% on 32 CPUs over  52.78 days; load 0.05 1/278 26069"  
[6] "briona: Idle time 99.51\% on 24 CPUs over  52.78 days; load 0.05 1/230 11857"  
[7] "cetona: Idle time 98.84\% on  8 CPUs over  21.18 days; load 0.05 1/130 31726"  
[8] "esaona: Idle time 37.47\% on 32 CPUs over  52.78 days; load 32.11 33/326 15939"
[9] "esbona: Idle time 85.73\% on 32 CPUs over  52.78 days; load 0.05 1/313 31753"  
[10] "escona: Idle time 84.79\% on 32 CPUs over  52.78 days; load 0.05 1/290 27626"  
[11] "esdona: Idle time 54.08\% on 32 CPUs over  52.78 days; load 32.13 33/327 7840" 
[12] "gemona: Idle time100.00\% on 24 CPUs over  52.78 days; load 0.05 1/230 14208"  
[13] "oraona: Idle time 83.99\% on 32 CPUs over  52.78 days; load 14.60 14/702 15421"
[14] "orbona: Idle time 84.47\% on 32 CPUs over  52.78 days; load 17.31 15/871 6256" 
[15] "orcona: Idle time 87.05\% on 32 CPUs over  52.78 days; load 14.53 14/701 19117"
[16] "ordona: Idle time 88.29\% on 32 CPUs over  52.78 days; load 14.49 14/702 28129"
[17] "ossona: Idle time 99.61\% on  8 CPUs over  52.78 days; load 0.05 1/130 9903"   
[18] "parona: Idle time 99.82\% on  8 CPUs over  52.78 days; load 0.05 1/130 10026"  
[19] "savona: Idle time 39.40\% on 32 CPUs over  52.78 days; load 15.97 5/1132 13169"
[20] "verona: Idle time 59.92\% on 48 CPUs over  52.79 days; load 24.56 26/775 12182"
}

